compatibility: remove example_base
Increase readability by removing unnecessary parent class.
